RSSEC-251 feat: add separated promo total sum if promo code was applyed
