+++ src/pages/basket/listeners/applyPromoCode.ts
@@ -1,4 +1,6 @@
+import { CartGetManager } from '../../../server/cart/getCartById';
 import { recalculatePromo } from './changeQuantity/recalculatePromo';
+import { recalculateValues } from './changeQuantity/recalculateValues';
 
 export function applyPromoCode() {
     document.addEventListener('click', async (event) => {
@@ -8,6 +10,12 @@ export function applyPromoCode() {
             const input = document.getElementById(`${id}-promo-input`) as HTMLInputElement;
             const price = document.getElementById(`${id}-cart-prod-price`) as HTMLDivElement;
             const prodName = document.getElementById(`${id}-cart-prod-name`) as HTMLDivElement;
+            // const totalSumBlock = document.getElementById('cart-sum') as HTMLDivElement;
+            const promoSumBlock = document.getElementById('cart-promo-sum') as HTMLDivElement;
+            const cartId = sessionStorage.getItem('newCartId') as string;
+            const myCart = new CartGetManager();
+            const cart = await myCart.getCartById(cartId);
+            console.log(cart);
             if (target.innerHTML === 'Apply') {
                 if (input.value === 'promo-laptop') {
                     input.remove();
@@ -20,6 +28,7 @@ export function applyPromoCode() {
                     recalculatePromo(id);
                 }
             } else {
+                promoSumBlock.innerHTML = '';
                 if (target.parentElement) target.parentElement.remove();
                 prodName.innerHTML += `<div id="lap-promo">
                                         <input id="${id}-promo-input" type="text" placeholder="PROMO">
@@ -27,7 +36,8 @@ export function applyPromoCode() {
                                     </div>`;
                 price.innerHTML = sessionStorage.getItem(id) as string;
                 sessionStorage.removeItem(id);
-                recalculatePromo(id);
+                recalculateValues(id, cart?.totalPrice.centAmount);
+                // recalculatePromo(id);
             }
         }
     });
