+++ src/pages/basket/listeners/changeQuantity/plusMinusOneProduct.ts
@@ -1,8 +1,7 @@
 import { decreaseItemInBasket, increaseItemInBasket } from '../../../../server/cart/changeQuantity';
+import { recalculate } from '../../functions/recalculate';
 import { minusOne } from './minusOne';
 import { plusOne } from './plusOne';
-import { recalculatePromo } from './recalculatePromo';
-import { recalculateValues } from './recalculateValues';
 
 export function plusMinusOneProduct(): void {
     const count = 1;
@@ -13,35 +12,13 @@ export function plusMinusOneProduct(): void {
             plusOne(target, count);
             const cart = await increaseItemInBasket(id);
 
-            let promoLapTrue = false;
-            (() => {
-                const allPromoBut = document.querySelectorAll('.lap-promo-but') as NodeList;
-                allPromoBut.forEach((but) => {
-                    const button = but as HTMLButtonElement;
-                    if (button.innerHTML === 'Special price DELETE') {
-                        promoLapTrue = true;
-                        return;
-                    }
-                });
-            })();
-            promoLapTrue ? recalculatePromo(id) : recalculateValues(id, cart?.totalPrice.centAmount);
+            recalculate(id, cart?.totalPrice.centAmount);
         } else if (target.className === 'minus-button') {
             const id = target.id.replace('-cart-prod-minus', '');
             minusOne(target, count);
             const cart = await decreaseItemInBasket(id);
 
-            let promoLapTrue = false;
-            (() => {
-                const allPromoBut = document.querySelectorAll('.lap-promo-but') as NodeList;
-                allPromoBut.forEach((but) => {
-                    const button = but as HTMLButtonElement;
-                    if (button.innerHTML === 'Special price DELETE') {
-                        promoLapTrue = true;
-                        return;
-                    }
-                });
-            })();
-            promoLapTrue ? recalculatePromo(id) : recalculateValues(id, cart?.totalPrice.centAmount);
+            recalculate(id, cart?.totalPrice.centAmount);
         }
     });
 }
