+++ src/pages/basket/listeners/removeProductFromCart.ts
@@ -14,21 +14,12 @@ export function removeProductFromCart(): void {
             const totalSumBlock = document.getElementById('cart-sum') as HTMLDivElement;
             totalSumBlock.innerHTML = String(addUpAllTheSums().toFixed(2));
 
-            const promoSum = document.getElementById('promo-total-sum') as HTMLDivElement;
-            if (promoSum && +totalSumBlock.innerHTML <= 200) promoSum.remove();
-
-            if (+totalSumBlock.innerHTML <= 200) {
-                const promoSum = document.getElementById('promo-total-sum') as HTMLDivElement;
-                if (promoSum) promoSum.remove();
-            }
-
             const header = document.getElementById('cart-header') as HTMLDivElement;
             const prodRows = document.querySelectorAll('.cart-prod-wrap') as NodeList;
             if (prodRows.length === 0) {
                 header.after(addEmptyInfo());
                 const clearBut = document.getElementById('cart-cleaner') as HTMLButtonElement;
                 clearBut.disabled = true;
-                if (promoSum) promoSum.remove();
                 showToast('All products were removed from basket');
             } else {
                 showToast('Product removed from basket');
