+++ src/pages/basket/utils/createPromoWindow.utils.ts
@@ -7,17 +7,12 @@ export async function createPromoWindow(): Promise<HTMLDivElement> {
     const promo = new PromoCodes();
     const promoCodes = (await promo.getPromoCodes()) as DiscountCode[];
     const firstPromo: string = promoCodes[1].code;
-    const secondPromo: string = promoCodes[0].code;
+    // const secondPromo: string = promoCodes[0].code;
     wrapper.innerHTML = `!!!Don't miss the chance!!!
                         <div id="lap-text">PROMO CODE FOR BUYING THE LAPTOP</div>
                         <div id="laptop">
                             <input id="inputText-lap" class="inputText" type="text" value=${firstPromo}>
                             <button id="copyText-lap" class="copyText">Copy</button>
-                        </div>
-                        <div id="sum-text">PROMO CODE FOR A LARGE PURCHASE (more than 200 GBP)</div>
-                        <div id="big-sum">
-                            <input id="inputText-sum" class="inputText" type="text" value=${secondPromo}>
-                            <button id="copyText-sum" class="copyText">Copy</button>
                         </div>`;
     return wrapper;
 }
